+++ R/cff_write.R
@@ -15,8 +15,8 @@
 #'   [cff_write_citation()].
 #' @param verbose Logical `TRUE/FALSE`. When `TRUE`, the function displays
 #'   informative messages.
-#' @param validate Logical `TRUE/FALSE`. Should the new file be validated
-#'   using `cff_validate()`?
+#' @param validate Logical `TRUE/FALSE`. Whether to validate the new file with
+#'   [cff_validate()].
 #' @param encoding The name of the encoding to be assumed. Default is `"UTF-8"`,
 #'   but it can be any other value as accepted by [iconv()], such as
 #'   `"ASCII//TRANSLIT"`.
@@ -72,11 +72,7 @@ cff_write <- function(
     x <- getwd()
   }
 
-  # Issue 86: remove "inst/CITATION" first.
-  fpath <- "./inst/CITATION"
-  if (r_citation) {
-    unlink(fpath)
-  }
+  cff_remove_existing_r_citation(r_citation)
 
   citat <- cff_create(
     x,
@@ -87,55 +83,9 @@ cff_write <- function(
     authors_roles = authors_roles
   )
 
-  # Ensure the output file uses a `.cff` extension.
-  if (!is_substring(outfile, ".cff$")) {
-    outfile <- paste0(outfile, ".cff")
-  }
-
-  # Create the directory if it does not exist.
-  outdir <- dirname(outfile)
-
-  if (!dir.exists(outdir)) {
-    dir.create(outdir, showWarnings = FALSE, recursive = TRUE)
-  }
-
-  # See https://github.com/r-universe-org/help/issues/382
-  # Write `CITATION.cff` with a header comment.
-  com <- c(
-    "# ------------------------------------------------",
-    "# CITATION.cff file created with {cffr} R package",
-    "# See also: https://docs.ropensci.org/cffr/",
-    "# ------------------------------------------------",
-    " "
-  )
-
-  # Make the best effort to get the encoding right.
-  com <- enc2utf8(com)
-  out_yaml <- enc2utf8(capture.output(print(citat)))
-  full_text <- enc2utf8(c(com, out_yaml, ""))
-  fh <- file(outfile, encoding = encoding)
-  on.exit(if (isOpen(fh)) close(fh))
-  writeLines(full_text, fh)
-
-  if (verbose) {
-    cli::cli_alert_success("{.file {outfile}} generated.")
-  }
-
-  # Add `CITATION.cff` to `.Rbuildignore`.
-  if (!is_cff(x) && x == getwd() && file_exist_abort(".Rbuildignore")) {
-    ignore <- readLines(".Rbuildignore")
-
-    # If not already present.
-    if (!("^CITATION\\.cff$" %in% ignore)) {
-      ignore <- c(ignore, "^CITATION\\.cff$")
-      ignore <- unique(ignore)
-
-      if (verbose) {
-        cli::cli_alert_info("Adding {.val {outfile}} to {.file .Rbuildignore}.")
-      }
-      writeLines(ignore, ".Rbuildignore")
-    }
-  }
+  outfile <- cff_normalize_outfile(outfile)
+  cff_write_file(citat, outfile, encoding, verbose)
+  cff_update_rbuildignore(x, outfile, verbose)
 
   if (validate) {
     cff_validate(outfile, verbose)
@@ -145,6 +95,85 @@ cff_write <- function(
   auto_r_citation(r_citation = r_citation, outfile = outfile, verbose = verbose)
 
   invisible(citat)
+}
+
+cff_remove_existing_r_citation <- function(r_citation) {
+  # Issue 86: remove "inst/CITATION" first.
+  if (r_citation) {
+    unlink("./inst/CITATION")
+  }
+}
+
+cff_normalize_outfile <- function(outfile) {
+  # Ensure the output file uses a `.cff` extension.
+  if (!is_substring(outfile, ".cff$")) {
+    outfile <- paste0(outfile, ".cff")
+  }
+
+  outfile
+}
+
+cff_write_file <- function(citat, outfile, encoding, verbose) {
+  cff_ensure_output_dir(outfile)
+
+  full_text <- cff_output_lines(citat)
+  fh <- file(outfile, encoding = encoding)
+  on.exit(if (isOpen(fh)) close(fh))
+  writeLines(full_text, fh)
+
+  if (verbose) {
+    cli::cli_alert_success("{.file {outfile}} generated.")
+  }
+}
+
+cff_ensure_output_dir <- function(outfile) {
+  outdir <- dirname(outfile)
+
+  if (!dir.exists(outdir)) {
+    dir.create(outdir, showWarnings = FALSE, recursive = TRUE)
+  }
+}
+
+cff_output_lines <- function(citat) {
+  # See https://github.com/r-universe-org/help/issues/382
+  # Write `CITATION.cff` with a header comment.
+  header <- c(
+    "# ------------------------------------------------",
+    "# CITATION.cff file created with {cffr} R package",
+    "# See also: https://docs.ropensci.org/cffr/",
+    "# ------------------------------------------------",
+    " "
+  )
+
+  # Make the best effort to get the encoding right.
+  header <- enc2utf8(header)
+  out_yaml <- enc2utf8(capture.output(print(citat)))
+  enc2utf8(c(header, out_yaml, ""))
+}
+
+cff_update_rbuildignore <- function(x, outfile, verbose) {
+  # Add `CITATION.cff` to `.Rbuildignore`.
+  if (is_cff(x) || x != getwd() || !file_exist_abort(".Rbuildignore")) {
+    return(invisible(NULL))
+  }
+
+  ignore <- readLines(".Rbuildignore")
+
+  # If not already present.
+  if ("^CITATION\\.cff$" %in% ignore) {
+    return(invisible(NULL))
+  }
+
+  ignore <- unique(c(ignore, "^CITATION\\.cff$"))
+
+  if (verbose) {
+    cli::cli_alert_info(
+      "Adding {.file {outfile}} to {.file .Rbuildignore}."
+    )
+  }
+  writeLines(ignore, ".Rbuildignore")
+
+  invisible(NULL)
 }
 
 auto_r_citation <- function(